# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: SuperAgent API
  version: 0.4.0
  description: >-
    Command, atomic Snapshot, queue mutation, and live-event contract for the
    durable AI Agent.
servers:
  - url: /
paths:
...
  /products/ai-agent/portal:
    get:
      operationId: getPortal
      summary: Return provider and tool choices for the launch portal
      responses:
        "200":
          description: Launch configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portal"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/start:
    post:
      operationId: startAgent
      summary: Start one durable Agent Flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAgentRequest"
      responses:
        "201":
          description: The Agent Flow was started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartAgentResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/messages:
    post:
      operationId: sendMessage
      summary: Queue a user message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "202":
          description: The command was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/questions/answer:
    post:
      operationId: answerQuestions
      summary: Answer the exact pending Agent input batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerQuestionsRequest"
      responses:
        "202":
          description: The answers were durably accepted and the input batch was closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/snapshot:
    get:
      operationId: getAgentSnapshot
      summary: Read one atomic durable Agent application view
      parameters:
        - name: flowId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FlowId"
      responses:
        "200":
          description: One atomic durable application Snapshot.
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSnapshot"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/archived-messages:
    get:
      operationId: getArchivedMessages
      summary: Read one immutable archived ten-message chunk
      parameters:
        - name: flowId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FlowId"
        - name: beforeSequence
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Sequence"
      responses:
        "200":
          description: One immutable history continuation before the boundary.
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                enum:
                  - no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/interaction-status:
    get:
      operationId: waitForAgentInteractionStatus
      summary: Wait for one durable Agent interaction status
      parameters:
        - name: flowId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FlowId"
        - name: expectedStatus
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AgentInteractionStatus"
      responses:
        "200":
          description: The expected durable interaction status was observed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInteractionState"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "504":
          description: The expected status was not observed before the bounded poll expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollTimeout"
  /products/ai-agent/message-queue/delete:
    post:
      operationId: deleteQueuedMessage
      summary: Delete one exact pending queued user message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueMutationRequest"
      responses:
        "200":
          description: The queued message was deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueMutationResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/message-queue/steer:
    post:
      operationId: steerQueuedMessage
      summary: Atomically move one queued user message into steering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueMutationRequest"
      responses:
        "200":
          description: The queued message was moved into steering.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueMutationResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/plans/execute:
    post:
      operationId: executePlan
      summary: Execute one exact durable plan revision
      description: >-
        Accepted only when the Agent is waiting for a message, the requested
        Plan revision is current, and no pending input, approval, timer, queued
        message, steering, or execution request blocks it. A conflicting
        boundary or stale revision returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutePlanRequest"
      responses:
        "202":
          description: The plan execution was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /products/ai-agent/tool-approvals:
    post:
      operationId: approveTool
      summary: Resolve one exact pending tool approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolApprovalRequest"
      responses:
        "202":
          description: The approval decision was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
...
components:
  responses:
    Problem:
      description: The request could not be completed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    FlowId:
      type: string
      minLength: 1
      maxLength: 255
      x-ogen-name: FlowID
...
    Portal:
      type: object
      additionalProperties: false
      required:
        - providers
        - mcpServers
        - tools
        - builtInTools
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/PortalProvider"
        mcpServers:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            $ref: "#/components/schemas/PortalTool"
        builtInTools:
          type: array
          items:
            $ref: "#/components/schemas/ToolName"
...
    StartAgentRequest:
      type: object
      additionalProperties: false
      required:
        - flowId
        - provider
        - model
        - systemPrompt
        - maxContextTokens
        - messageRetentionLimit
        - mcpEnabled
        - enabledMcpServers
        - enabledTools
      properties:
        flowId:
          $ref: "#/components/schemas/FlowId"
        provider:
          $ref: "#/components/schemas/Provider"
        model:
          type: string
          minLength: 1
          maxLength: 255
        systemPrompt:
          type: string
          minLength: 1
          maxLength: 100000
        compactionModel:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
        maxContextTokens:
          type: integer
          minimum: 1
          maximum: 2000000
        compactionTriggerFraction:
          type: number
          format: double
          minimum: 0
          maximum: 1
          exclusiveMinimum: true
          exclusiveMaximum: true
          default: 0.85
        compactionKeepFraction:
          type: number
          format: double
          minimum: 0
          maximum: 1
          exclusiveMinimum: true
          exclusiveMaximum: true
          default: 0.1
        messageRetentionLimit:
          type: integer
          minimum: 20
          maximum: 1000000
          multipleOf: 10
        mcpEnabled:
          type: boolean
        enabledMcpServers:
          type: array
          uniqueItems: true
          maxItems: 1000
          items:
            type: string
            minLength: 1
        enabledTools:
          type: array
          uniqueItems: true
          maxItems: 10000
          items:
            $ref: "#/components/schemas/ToolName"
...