# internal/api/generated/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn4AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn20AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn16AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn5AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn18AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'h': // Prefix: "healthz"

				if l := len("healthz"); len(elem) >= l && elem[0:l] == "healthz" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetHealthRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'p': // Prefix: "products/ai-agent/"

				if l := len("products/ai-agent/"); len(elem) >= l && elem[0:l] == "products/ai-agent/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "archived-messages"

					if l := len("archived-messages"); len(elem) >= l && elem[0:l] == "archived-messages" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetArchivedMessagesRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'e': // Prefix: "events"

					if l := len("events"); len(elem) >= l && elem[0:l] == "events" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleReadEventRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/recent"

						if l := len("/recent"); len(elem) >= l && elem[0:l] == "/recent" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleListRecentEventsRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'i': // Prefix: "interaction-status"

					if l := len("interaction-status"); len(elem) >= l && elem[0:l] == "interaction-status" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleWaitForAgentInteractionStatusRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: nil,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'm': // Prefix: "message"

					if l := len("message"); len(elem) >= l && elem[0:l] == "message" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '-': // Prefix: "-queue/"

						if l := len("-queue/"); len(elem) >= l && elem[0:l] == "-queue/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'd': // Prefix: "delete"

							if l := len("delete"); len(elem) >= l && elem[0:l] == "delete" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleDeleteQueuedMessageRequest([0]string{}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn4AllowedHeaders,
										acceptPost:     "application/json",
										acceptPatch:    "",
									})
								}

								return
							}

						case 's': // Prefix: "steer"

							if l := len("steer"); len(elem) >= l && elem[0:l] == "steer" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "POST":
									s.handleSteerQueuedMessageRequest([0]string{}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "POST",
										allowedHeaders: rn20AllowedHeaders,
										acceptPost:     "application/json",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					case 's': // Prefix: "s"

						if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleSendMessageRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn16AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'p': // Prefix: "p"

					if l := len("p"); len(elem) >= l && elem[0:l] == "p" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'l': // Prefix: "lans/execute"

						if l := len("lans/execute"); len(elem) >= l && elem[0:l] == "lans/execute" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleExecutePlanRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn5AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'o': // Prefix: "ortal"

						if l := len("ortal"); len(elem) >= l && elem[0:l] == "ortal" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetPortalRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 'q': // Prefix: "questions/answer"

					if l := len("questions/answer"); len(elem) >= l && elem[0:l] == "questions/answer" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleAnswerQuestionsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'n': // Prefix: "napshot"

						if l := len("napshot"); len(elem) >= l && elem[0:l] == "napshot" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetAgentSnapshotRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET",
									allowedHeaders: nil,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 't': // Prefix: "tart"

						if l := len("tart"); len(elem) >= l && elem[0:l] == "tart" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleStartAgentRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "POST",
									allowedHeaders: rn18AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				case 't': // Prefix: "tool-approvals"

					if l := len("tool-approvals"); len(elem) >= l && elem[0:l] == "tool-approvals" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleApproveToolRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn3AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'r': // Prefix: "readyz"

				if l := len("readyz"); len(elem) >= l && elem[0:l] == "readyz" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetReadinessRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'h': // Prefix: "healthz"

				if l := len("healthz"); len(elem) >= l && elem[0:l] == "healthz" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetHealthOperation
						r.summary = "Report process liveness"
						r.operationID = "getHealth"
						r.operationGroup = ""
						r.pathPattern = "/healthz"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'p': // Prefix: "products/ai-agent/"

				if l := len("products/ai-agent/"); len(elem) >= l && elem[0:l] == "products/ai-agent/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "archived-messages"

					if l := len("archived-messages"); len(elem) >= l && elem[0:l] == "archived-messages" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetArchivedMessagesOperation
							r.summary = "Read one immutable archived ten-message chunk"
							r.operationID = "getArchivedMessages"
							r.operationGroup = ""
							r.pathPattern = "/products/ai-agent/archived-messages"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'e': // Prefix: "events"

					if l := len("events"); len(elem) >= l && elem[0:l] == "events" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ReadEventOperation
							r.summary = "Long-poll one best-effort Agent Stream"
							r.operationID = "readEvent"
							r.operationGroup = ""
							r.pathPattern = "/products/ai-agent/events"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/recent"

						if l := len("/recent"); len(elem) >= l && elem[0:l] == "/recent" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = ListRecentEventsOperation
								r.summary = "List the configured recent tail of one best-effort Agent Stream"
								r.operationID = "listRecentEvents"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/events/recent"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'i': // Prefix: "interaction-status"

					if l := len("interaction-status"); len(elem) >= l && elem[0:l] == "interaction-status" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = WaitForAgentInteractionStatusOperation
							r.summary = "Wait for one durable Agent interaction status"
							r.operationID = "waitForAgentInteractionStatus"
							r.operationGroup = ""
							r.pathPattern = "/products/ai-agent/interaction-status"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'm': // Prefix: "message"

					if l := len("message"); len(elem) >= l && elem[0:l] == "message" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case '-': // Prefix: "-queue/"

						if l := len("-queue/"); len(elem) >= l && elem[0:l] == "-queue/" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 'd': // Prefix: "delete"

							if l := len("delete"); len(elem) >= l && elem[0:l] == "delete" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = DeleteQueuedMessageOperation
									r.summary = "Delete one exact pending queued user message"
									r.operationID = "deleteQueuedMessage"
									r.operationGroup = ""
									r.pathPattern = "/products/ai-agent/message-queue/delete"
									r.args = args
									r.count = 0
									return r, true
								default:
									return
								}
							}

						case 's': // Prefix: "steer"

							if l := len("steer"); len(elem) >= l && elem[0:l] == "steer" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "POST":
									r.name = SteerQueuedMessageOperation
									r.summary = "Atomically move one queued user message into steering"
									r.operationID = "steerQueuedMessage"
									r.operationGroup = ""
									r.pathPattern = "/products/ai-agent/message-queue/steer"
									r.args = args
									r.count = 0
									return r, true
								default:
									return
								}
							}

						}

					case 's': // Prefix: "s"

						if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = SendMessageOperation
								r.summary = "Queue a user message"
								r.operationID = "sendMessage"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/messages"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'p': // Prefix: "p"

					if l := len("p"); len(elem) >= l && elem[0:l] == "p" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'l': // Prefix: "lans/execute"

						if l := len("lans/execute"); len(elem) >= l && elem[0:l] == "lans/execute" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = ExecutePlanOperation
								r.summary = "Execute one exact durable plan revision"
								r.operationID = "executePlan"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/plans/execute"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'o': // Prefix: "ortal"

						if l := len("ortal"); len(elem) >= l && elem[0:l] == "ortal" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetPortalOperation
								r.summary = "Return provider and tool choices for the launch portal"
								r.operationID = "getPortal"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/portal"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'q': // Prefix: "questions/answer"

					if l := len("questions/answer"); len(elem) >= l && elem[0:l] == "questions/answer" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = AnswerQuestionsOperation
							r.summary = "Answer the exact pending Agent input batch"
							r.operationID = "answerQuestions"
							r.operationGroup = ""
							r.pathPattern = "/products/ai-agent/questions/answer"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 's': // Prefix: "s"

					if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'n': // Prefix: "napshot"

						if l := len("napshot"); len(elem) >= l && elem[0:l] == "napshot" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetAgentSnapshotOperation
								r.summary = "Read one atomic durable Agent application view"
								r.operationID = "getAgentSnapshot"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/snapshot"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 't': // Prefix: "tart"

						if l := len("tart"); len(elem) >= l && elem[0:l] == "tart" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "POST":
								r.name = StartAgentOperation
								r.summary = "Start one durable Agent Flow"
								r.operationID = "startAgent"
								r.operationGroup = ""
								r.pathPattern = "/products/ai-agent/start"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 't': // Prefix: "tool-approvals"

					if l := len("tool-approvals"); len(elem) >= l && elem[0:l] == "tool-approvals" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = ApproveToolOperation
							r.summary = "Resolve one exact pending tool approval"
							r.operationID = "approveTool"
							r.operationGroup = ""
							r.pathPattern = "/products/ai-agent/tool-approvals"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'r': // Prefix: "readyz"

				if l := len("readyz"); len(elem) >= l && elem[0:l] == "readyz" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetReadinessOperation
						r.summary = "Report dependency readiness"
						r.operationID = "getReadiness"
						r.operationGroup = ""
						r.pathPattern = "/readyz"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}